+++ h2integrate/core/supported_models.py
@@ -2,6 +2,7 @@ from h2integrate.resource.river import RiverResource
 from h2integrate.core.feedstocks import FeedstockCostModel, FeedstockPerformanceModel
 from h2integrate.transporters.pipe import PipePerformanceModel
 from h2integrate.transporters.cable import CablePerformanceModel
+from h2integrate.converters.grid.grid import GridCostModel, GridPerformanceModel
 from h2integrate.finances.profast_lco import ProFastLCO
 from h2integrate.finances.profast_npv import ProFastNPV
 from h2integrate.converters.steel.steel import SteelPerformanceModel, SteelCostAndFinancialModel
@@ -95,16 +96,17 @@ from h2integrate.resource.solar.nrel_developer_goes_api_models import (
 from h2integrate.converters.hydrogen.eco_tools_pem_electrolyzer import (
     ECOElectrolyzerPerformanceModel,
 )
-from h2integrate.control.control_strategies.openloop_controllers import (
-    DemandOpenLoopController,
-    PassThroughOpenLoopController,
-)
 from h2integrate.converters.water_power.hydro_plant_run_of_river import (
     RunOfRiverHydroCostModel,
     RunOfRiverHydroPerformanceModel,
 )
 from h2integrate.converters.hydrogen.geologic.simple_natural_geoh2 import (
     NaturalGeoH2PerformanceModel,
+)
+from h2integrate.resource.solar.nrel_developer_himawari_api_models import (
+    Himawari7SolarAPI,
+    Himawari8SolarAPI,
+    HimawariTMYSolarAPI,
 )
 from h2integrate.control.control_rules.converters.generic_converter import (
     PyomoDispatchGenericConverter,
@@ -127,6 +129,22 @@ from h2integrate.converters.hydrogen.geologic.templeton_serpentinization import 
 from h2integrate.control.control_rules.storage.pyomo_storage_rule_baseclass import (
     PyomoRuleStorageBaseclass,
 )
+from h2integrate.control.control_strategies.passthrough_openloop_controller import (
+    PassThroughOpenLoopController,
+)
+from h2integrate.resource.solar.nrel_developer_meteosat_prime_meridian_models import (
+    MeteosatPrimeMeridianSolarAPI,
+    MeteosatPrimeMeridianTMYSolarAPI,
+)
+from h2integrate.control.control_strategies.storage.demand_openloop_controller import (
+    DemandOpenLoopStorageController,
+)
+from h2integrate.control.control_strategies.converters.demand_openloop_controller import (
+    DemandOpenLoopConverterController,
+)
+from h2integrate.control.control_strategies.converters.flexible_demand_openloop_controller import (
+    FlexibleDemandOpenLoopConverterController,
+)
 
 
 supported_models = {
@@ -138,6 +156,11 @@ supported_models = {
     "goes_conus_solar_v4_api": GOESConusSolarAPI,
     "goes_fulldisc_solar_v4_api": GOESFullDiscSolarAPI,
     "goes_tmy_solar_v4_api": GOESTMYSolarAPI,
+    "meteosat_solar_v4_api": MeteosatPrimeMeridianSolarAPI,
+    "meteosat_tmy_solar_v4_api": MeteosatPrimeMeridianTMYSolarAPI,
+    "himawari7_solar_v3_api": Himawari7SolarAPI,
+    "himawari8_solar_v3_api": Himawari8SolarAPI,
+    "himawari_tmy_solar_v3_api": HimawariTMYSolarAPI,
     # Converters
     "atb_wind_cost": ATBWindPlantCostModel,
     "pysam_wind_plant_performance": PYSAMWindPlantPerformanceModel,
@@ -212,19 +235,31 @@ supported_models = {
     "simple_generic_storage": SimpleGenericStorage,
     # Control
     "pass_through_controller": PassThroughOpenLoopController,
-    "demand_open_loop_controller": DemandOpenLoopController,
+    "demand_open_loop_storage_controller": DemandOpenLoopStorageController,
     "heuristic_load_following_controller": HeuristicLoadFollowingController,
+    "demand_open_loop_converter_controller": DemandOpenLoopConverterController,
+    "flexible_demand_open_loop_converter_controller": FlexibleDemandOpenLoopConverterController,
     # Dispatch
     "pyomo_dispatch_generic_converter": PyomoDispatchGenericConverter,
     "pyomo_dispatch_generic_storage": PyomoRuleStorageBaseclass,
     # Feedstock
     "feedstock_performance": FeedstockPerformanceModel,
     "feedstock_cost": FeedstockCostModel,
+    # Grid
+    "grid_performance": GridPerformanceModel,
+    "grid_cost": GridCostModel,
     # Finance
     "ProFastComp": ProFastLCO,
     "ProFastNPV": ProFastNPV,
     "NumpyFinancialNPV": NumpyFinancialNPV,
 }
 
-
-electricity_producing_techs = ["wind", "solar", "pv", "river", "hopp", "natural_gas_plant"]
+electricity_producing_techs = [
+    "wind",
+    "solar",
+    "pv",
+    "river",
+    "hopp",
+    "natural_gas_plant",
+    "grid_buy",
+]