+++ CHANGELOG.md
@@ -31,6 +31,10 @@
 - Added ability for latitude and longitude to be design variables in design sweep
 - Move geologic hydrogen models into specific geoh2 subsurface converters
 - Added standalone iron mine performance and cost model
+- Added solar resource models for Meteosat Prime Meridian and Himawari datasets available through NSRDB
+- Improved readability of the postprocessing printout by simplifying numerical representation, especially for years
+- Added grid converter performance and cost model which can be used to buy, sell, or buy and sell electricity to/from the grid
+- Add open-loop load demand controllers: `DemandOpenLoopConverterController` and `FlexibleDemandOpenLoopConverterController`
 
 ## 0.4.0 [October 1, 2025]